Lexer: Added support for scienific notation to parse doubles Added can-parse: utilitary tool that parses and outputs to the console a database's content Parses BS_ (but doesn't do anything with it) Fix global comments where not correctly parsed (parsed but not handled) Added comment to CANFrame API

+++ include/Token.h
@@ -37,7 +37,7 @@ public:
   }
 
   double toDouble() const {
-    return std::stof(image_);
+    return std::stod(image_);
   }
   
 private:
